requirements.txt
selenium~=4.23.0
python-dotenv~=1.0.1
pytest~=8.3.1